+++ builders/changesetAction/tsconfig.json
@@ -1,8 +1,11 @@
 {
   "extends": "tsconfig/tsconfig.base.json",
   "compilerOptions": {
-    "target": "es2015",
-    "moduleResolution": "node"
+    "target": "es6",                          /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', 'ES2018', 'ES2019' or 'ESNEXT'. */
+    "module": "commonjs",                     /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */
+    "outDir": "./lib",                        /* Redirect output structure to the directory. */
+    "rootDir": "./src",                       /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
   },
-  "include": ["src"]
+  "include": ["src"],
+  "exclude": ["node_modules", "playground"]
 }
